Refactor classes to separate files
