+++ rtCommon/bidsIncremental.py
@@ -2,383 +2,278 @@
 
 bidsIncremental.py
 
-This script includes all of the functions that are needed to convert between
-DICOM, BIDS-Incremental (BIDS-I), and BIDS formats.
+Implements the BIDS Incremental data type used for streaming BIDS data between
+different applications.
 
 -----------------------------------------------------------------------------"""
-
-from bids.layout import parse_file_entities
-
-import json
 import logging
-import nibabel as nib
-import numpy as np
 import os
-import pydicom
+import re
+from typing import List
 
-from rtCommon.bidsStructures import BidsArchive, BidsIncremental
-import rtCommon.bidsUtils as bidsUtils
+import nibabel as nib
+
 from rtCommon.errors import ValidationError
-from rtCommon.imageHandling import convertDicomImgToNifti, readNifti
+from rtCommon.bidsCommon import (
+    BidsFileExtension,
+    BIDS_VERSION,
+    DATASET_DESC_REQ_FIELDS,
+)
 
 logger = logging.getLogger(__name__)
 
-# Set to True to skip validating Nifti header when appending
-DISABLE_NIFTI_HEADER_CHECK = False
 
-# Set to True to skip validating metadata when appending
-DISABLE_METADATA_CHECK = True
-
-
-def getMetadata(dicomImg: pydicom.dataset.Dataset) -> (dict, dict):
+class BidsIncremental:
     """
-    Returns the public and private metadata from the provided DICOM image.
-
-    Args:
-        dicomImg: A pydicom object to read metadata from.
-    Returns:
-        Tuple of 2 dictionaries, the first containing the public metadata from
-        the image and the second containing the private metadata.
+    BIDS Incremental data format suitable for streaming BIDS Archives
     """
-    if not isinstance(dicomImg, pydicom.dataset.Dataset):
-        raise ValidationError("Expected pydicom.dataset.Dataset as argument")
+    def __init__(self,
+                 niftiImg: nib.Nifti1Image,
+                 imgMetadata: dict,
+                 datasetMetadata: dict = None):
+        """
+        Initializes a BIDS Incremental object with required data and metadata.
 
-    publicMeta = {}
-    privateMeta = {}
+        Args:
+            niftiImg: Nifti image for this BIDS-I as an NiBabel NiftiImage.
+            imgMetadata: Required BIDS metadata for the NIfTI filename. Required
+                fields: 'sub', 'task', 'contrast_type'.
+            datasetMetadata: Top-level dataset metadata for the BIDS dataset
+                this represents.
 
-    ignoredTags = ['Pixel Data']
+        Raises:
+            ValidationError: If any required metadata field is not provided.
 
-    for elem in dicomImg:
-        if elem.name in ignoredTags:
-            continue
-
-        cleanedKey = bidsUtils.makeDicomFieldBidsCompatible(elem.name)
-        # in DICOM, public tags have even group numbers and private tags are odd
-        # http://dicom.nema.org/dicom/2013/output/chtml/part05/chapter_7.html
-        if elem.tag.is_private:
-            privateMeta[cleanedKey] = str(elem.value)
+        """
+        if niftiImg is None:
+            raise ValidationError("Image cannot be None!")
+        elif not (isinstance(niftiImg, nib.Nifti1Image) or
+                  isinstance(niftiImg, nib.Nifti2Image)):
+            raise ValidationError("Image must be NIBabel Nifti 1 or 2 image!")
         else:
-            publicMeta[cleanedKey] = str(elem.value)
+            self.image = niftiImg
+            self.header = niftiImg.header
 
-    return (publicMeta, privateMeta)
+        missingMetadata = self.findMissingMetadata(imgMetadata)
+        if missingMetadata != []:
+            errorMsg = "Image metadata missing these required fields: {}" \
+                .format(missingMetadata)
+            raise ValidationError(errorMsg)
 
+        self.imgMetadata = imgMetadata
 
-def dicomToBidsinc(dicomImg: pydicom.dataset.Dataset) -> BidsIncremental:
-    # TODO(spolcyn): Do this all in memory -- dicom2nifti is promising
-    # Put extra metadata in sidecar JSON file
-    #
-    # NOTE: This is not the final version of this method.
-    # The conversion from DICOM to BIDS-I and gathering all required metadata
-    # can be complex, as DICOM doesn't necessarily have the metadata required
-    # for BIDS in it by default. Thus, another component will handle the logic
-    # and error handling surrounding this.
-    niftiImage = convertDicomImgToNifti(dicomImg)
-    logger.debug("Nifti header after conversion is: %s", niftiImage.header)
-    publicMeta, privateMeta = getMetadata(dicomImg)
+        if datasetMetadata is None:
+            datasetMetadata = {}
+            datasetMetadata["Name"] = "BIDS-Incremental Dataset"
+            datasetMetadata["BIDSVersion"] = str(BIDS_VERSION)
+        else:
+            missingFields = [field for
+                             field in DATASET_DESC_REQ_FIELDS
+                             if datasetMetadata.get(field) is None]
 
-    publicMeta.update(privateMeta)  # combine metadata dictionaries
-    requiredMetadata = {'sub': '002', 'task': 'story', 'contrast_label': 'bold'}
-    publicMeta.update(requiredMetadata)
-    return BidsIncremental(niftiImage, publicMeta)
+            if missingFields != []:
+                errorMsg = "Dataset description provided, but missing these \
+                        required fields: " + missingFields
+                raise ValidationError(missingFields)
 
+        self.version = 1
 
-def verifyNiftiHeadersMatch(img1: nib.Nifti1Image, img2: nib.Nifti1Image):
-    """
-    Verifies that two Nifti image headers match in along a defined set of
-    NIfTI header fields which should not change during a continuous fMRI
-    scanning session.
+    def __str__(self):
+        return "Image shape: {}; # Metadata Keys: {}; Version: {}".format(
+            self.image.get_fdata().shape,
+            len(self.imgMetadata.keys()),
+            self.version)
 
-    This is primarily intended as a safety check, and does not conclusively
-    determine that two images are valid to append to together or are part of the
-    same scanning session.
+    @staticmethod
+    def findMissingMetadata(metadata: dict) -> List[str]:
+        requiredMetadata = ['sub', 'task', 'contrast_label']
+        if metadata is None:
+            return requiredMetadata
+        else:
+            return [key for key in requiredMetadata
+                    if metadata.get(key) is None]
 
-    Args:
-        header1: First Nifti header to compare (dict of numpy arrays)
-        header2: Second Nifti header to compare (dict of numpy arrays)
+    def getFieldLabelString(self, bidsField: str) -> str:
+        """
+        Extracts the field-label combination for the provided BIDS Standard
+        field from this BIDS Incremental. Valid fields are defined in the "Task
+        (including resting state) imaging data" section of the BIDS standard,
+        and include examples like "sub", "ses", "task", and "run".
 
-    Returns:
-        True if the headers match along the required dimensions, false
-        otherwise.
+        Returns:
+            The label (e.g, '01' for 'sub') if present, None otherwise.
+        """
+        # See if value already cached
+        label = self.imgMetadata.get(bidsField, None)
+        if label is not None:
+            return label
 
-    """
-    fieldsToMatch = ["intent_p1", "intent_p2", "intent_p3", "intent_code",
-                     "dim_info", "datatype", "bitpix", "xyzt_units",
-                     "slice_duration", "toffset", "scl_slope", "scl_inter",
-                     "qform_code", "quatern_b", "quatern_c", "quatern_d",
-                     "qoffset_x", "qoffset_y", "qoffset_z",
-                     "sform_code", "srow_x", "srow_y", "srow_z"]
+        # Attempt to extract from protocol name metadata if present
+        protocolName = self.imgMetadata.get('ProtocolName', None)
+        if protocolName is None:
+            return None
 
-    header1 = img1.header
-    header2 = img2.header
+        prefix = "(?:(?<=_)|(?<=^))"  # match beginning of string or underscore
+        suffix = "(?:(?=_)|(?=$))"  # match end of string or underscore
 
-    for field in fieldsToMatch:
-        v1 = header1.get(field)
-        v2 = header2.get(field)
+        pattern = "{prefix}(?:{field}-)(.+?){suffix}".format(prefix=prefix,
+                                                             field=bidsField,
+                                                             suffix=suffix)
+        result = re.search(pattern, protocolName)
 
-        # Use slightly more complicated check to properly match nan values
-        if not (np.allclose(v1, v2, atol=0.0, equal_nan=True)):
-            logger.debug("Nifti headers don't match on field: %s \
-                         (v1: %s, v2: %s)\n", field, v1, v2)
-            if DISABLE_NIFTI_HEADER_CHECK:
-                continue
-            else:
-                return False
+        if len(result.groups()) == 1:
+            return result.group(1)
+        else:
+            logger.debug("Failed to find exactly one match in protocol name \
+                    \'%s\' for field %s", protocolName, bidsField)
+            return None
 
-    # For pixel dimensions, 0 and 1 are equivalent -- any value in a higher
-    # index than the number of dimensions specified in the 'dim' field will be
-    # ignored, and a 0 in a non-ignored index makes no sense
-    field = "pixdim"
-    v1 = header1.get(field)
-    v2 = header2.get(field)
-    v1 = np.where(v1 == 0, 1, v1)
-    v2 = np.where(v2 == 0, 1, v2)
+    def getSubjectID(self):
+        return self.getFieldLabelString('sub')
 
-    if not (np.allclose(v1, v2, atol=0.0, equal_nan=True)):
-        logger.debug("Nifti headers don't match on field: %s \
-                     (v1: %s, v2: %s)\n", field, v1, v2)
-        if DISABLE_NIFTI_HEADER_CHECK:
+    def getSessionName(self):
+        return self.getFieldLabelString('ses')
+
+    def getTaskName(self):
+        return self.getFieldLabelString('task')
+
+    def getDataTypeName(self):
+        """ func or anat """
+        return "func"
+
+    def getRunLabel(self):
+        return self.getFieldLabelString('run')
+
+    def getContrastLabel(self):
+        return self.getFieldLabelString('contrast_label')
+
+    def makeBidsFileName(self, extension: BidsFileExtension) -> str:
+        """
+        Create the a BIDS-compatible file name based on the metadata. General
+        format of the filename, per BIDS standard 1.4.1, is as follows (items in
+        square brackets [] are considered optional):
+
+        sub-<label>[_ses-<label>]_task-<label>[_acq-<label>] [_ce-<label>]
+        [_dir-<label>][_rec-<label>][_run-<index>]
+        [_echo-<index>]_<contrast_label >.ext
+
+        Args:
+            extension: The extension for the file, e.g., 'nii' for images or
+                'json' for metadata
+
+        Return:
+            Filename from metadata according to BIDS standard 1.4.1.
+        """
+        labelPairs = []  # all potential BIDS field-label pairs in the filename
+
+        labelPairs.append('sub-' + self.getSubjectID())
+
+        sesName = self.getSessionName()
+        if sesName:
+            labelPairs.append('ses-' + sesName)
+
+        labelPairs.append('task-' + self.getTaskName())
+
+        runName = self.getRunLabel()
+        if runName:
+            labelPairs.append('run-' + runName)
+
+        labelPairs.append(self.getContrastLabel())
+
+        """
+        # distinguish using diff params for acquiring same task
+        acqLabel = getAcqLabel()
+        # "distinguish sequences using different constrast enhanced images"
+        ceLabel = getCeLabel()
+        # "distinguish different phase-encoding directions"
+        dirLabel = getDirLabel()
+        # "distinguish different...reconstruction algorithms"
+        recLabel = getRecLabel()
+        # "more than one run of same task"
+        runLabel = getRunLabel()
+        # "multi echo data"
+        echoLabel = getEchoLabel()
+        """
+
+        return '_'.join(labelPairs) + extension.value
+
+    def getImageFileName(self) -> str:
+        return self.makeBidsFileName(BidsFileExtension.IMAGE)
+
+    def getMetadataFileName(self) -> str:
+        return self.makeBidsFileName(BidsFileExtension.METADATA)
+
+    def getDatasetName(self) -> str:
+        return 'dataset'
+
+    def makeDataDirPath(self) -> str:
+        """
+        Returns the path to the data directory this incremental's data would be
+        in if it were in a full BIDS archive.
+
+        Returns:
+            Path string relative to root of the imaginary dataset.
+        Examples:
+            >>> print(bidsi.makePath())
+            /sub-01/ses-2011/anat/
+        """
+        return os.path.join("",
+                            'sub-' + self.getSubjectID(),
+                            'ses-' + self.getSessionName(),
+                            self.getDataTypeName(),
+                            "")
+
+    def makeImageFilePath(self) -> str:
+        dataDir = self.makeDataDirPath()
+        return os.path.join(dataDir, self.getImageFileName())
+
+    def writeToFile(self, directoryPath: str):
+        # Create folder structure -- just func for now
+        datasetDir = os.path.join(directoryPath, "dataset")
+        try:
+            os.mkdir(datasetDir)
+        except FileExistsError:
             pass
+
+        subjectName = self.getSubjectID()
+        subjectDir = os.path.join(datasetDir, "sub-" + subjectName)
+        try:
+            os.mkdir(subjectDir)
+        except FileExistsError:
+            pass
+
+        # get session/task info for directories
+        session = self.getSessionName()
+        if session:
+            session = session.group(0)
+            sessionDir = os.path.join(subjectDir, session)
+            try:
+                os.mkdir(sessionDir)
+            except FileExistsError:
+                pass
+
+        task = self.getTaskName()
+        if task:
+            task = task.group(0)
+
+        if session:
+            funcDir = os.path.join(sessionDir, "func")
         else:
-            return False
+            funcDir = os.path.join(subjectDir, "func")
 
-    return True
+        try:
+            os.mkdir(funcDir)
+        except FileExistsError:
+            pass
 
+        # Write out nifti to func folder
+        niftiFilename = os.path.join(funcDir, 'sub-{}'.format(subjectName))
+        if session:
+            niftiFilename += '_' + task
+        print(niftiFilename)
 
-def verifyMetadataMatch(meta1: dict, meta2: dict):
-    """
-    Verifies two metadata dictionaries match in a set of required fields. If a
-    field is present in only one or neither of the two dictionaries, this is
-    considered a match.
+        nib.save(self.image, os.path.join(funcDir, niftiFilename))
 
-    This is primarily intended as a safety check, and does not conclusively
-    determine that two images are valid to append to together or are part of the
-    same series.
+        # Write out metadata
 
-    Args:
-        meta1: First metadata dictionary
-        meta2: Second metadata dictionary
-
-    Returns:
-        True if all keys that are present in both dictionaries have equivalent
-        values, False otherwise.
-
-    """
-    if DISABLE_METADATA_CHECK:
-        return True
-
-    matchFields = ["Modality", "MagneticFieldStrength", "ImagingFrequency",
-                   "Manufacturer", "ManufacturersModelName", "InstitutionName",
-                   "InstitutionAddress", "DeviceSerialNumber", "StationName",
-                   "BodyPartExamined", "PatientPosition", "EchoTime",
-                   "ProcedureStepDescription", "SoftwareVersions",
-                   "MRAcquisitionType", "SeriesDescription", "ProtocolName",
-                   "ScanningSequence", "SequenceVariant", "ScanOptions",
-                   "SequenceName", "SpacingBetweenSlices", "SliceThickness",
-                   "ImageType", "RepetitionTime", "PhaseEncodingDirection",
-                   "FlipAngle", "InPlanePhaseEncodingDirectionDICOM",
-                   "ImageOrientationPatientDICOM", "PartialFourier"]
-
-    # If either field is None, short-circuit and continue checking other fields
-    for field in matchFields:
-        field1 = meta1.get(field)
-        if field1 is None:
-            continue
-
-        field2 = meta2.get(field)
-        if field2 is None:
-            continue
-
-        if field1 != field2:
-            logger.debug("Metadata doen't match on field: %s \
-                         (v1: %s, v2: %s)\n", field, field1, field2)
-            return False
-
-    # These fields should not match between two images for a valid append
-    differentFields = ["AcquisitionTime", "Acquisition Number"]
-
-    for field in differentFields:
-        logger.debug("Verifying: %s", field)
-        field1 = meta1.get(field)
-        if field1 is None:
-            continue
-
-        field2 = meta2.get(field)
-        if field2 is None:
-            continue
-
-        if field1 == field2:
-            logger.debug("Metadata matches (shouldn't) on field: %s \
-                         (v1: %s, v2: %s)\n", field, field1, field2)
-            return False
-
-    return True
-
-
-def appendBidsinc(incremental: BidsIncremental,
-                  archive: BidsArchive,
-                  makePath: bool = False) -> None:
-    """
-    Appends the provided BIDS Incremental imaging and metadata to the provided
-    BIDS archive. By default, expects that the incremental represents a valid
-    subset of the archive and no additional directory paths will need to be
-    created within the archive (this behavior can be overriden).
-
-    Args:
-        incremental: BIDS Incremental file containing image data and metadata
-        archive: BIDS Archive file to append image data and metadata to
-        makePath: Create the directory path for the BIDS-I in the archive if it
-            doesn't already exist.
-
-    Returns:
-        None
-
-    Raises:
-        ValidationError: If the image path within the BIDS incremental does not
-            match any existing paths within the archive, and no override is set
-
-    """
-    # 1) Create target path for image in archive
-    imgDirPath = incremental.makeDataDirPath()
-    imgPath = os.path.join(imgDirPath, incremental.getImageFileName())
-    metadataPath = os.path.join(imgDirPath, incremental.getMetadataFileName())
-
-    # 2) Verify we have a valid way to append the image to the archive. 3 cases:
-    # 2.1) Image already exists within archive, append this Nifti to that Nifti
-    # 2.2) Image doesn't exist in archive, but rest of the path is valid for the
-    # archive; create new Nifti file within the archive
-    # 2.3) Neither image nor path is valid for provided archive; fail append
-    if archive.pathExists(imgPath):
-        logger.debug("Image exists in archive, appending")
-        archiveImg = archive.getImage(imgPath)
-
-        # Validate header match
-        if not verifyNiftiHeadersMatch(incremental.image,
-                                       archiveImg):
-            raise ValidationError("Nifti headers failed validation!")
-        if not verifyMetadataMatch(incremental.imgMetadata,
-                                   archive.getMetadata(metadataPath)):
-            raise ValidationError("Image metadata failed validation!")
-
-        # Build 4-D NIfTI if archive has 3-D, concat to 4-D otherwise
-        incrementalData = incremental.image.get_fdata()
-        archiveData = archiveImg.get_fdata()
-
-        if len(archiveData.shape) == 3:
-            newArchiveData = np.stack((archiveData, incrementalData), axis=3)
-        else:
-            incrementalData = np.expand_dims(incrementalData, 3)
-            newArchiveData = np.concatenate((archiveData, incrementalData),
-                                            axis=3)
-
-        newImg = nib.Nifti1Image(newArchiveData,
-                                 archiveImg.affine,
-                                 header=archiveImg.header)
-        newImg.update_header()
-        archive.addImage(newImg, imgPath)
-
-    elif archive.pathExists(imgDirPath) or makePath is True:
-        logger.debug("Image doesn't exist in archive, creating")
-        archive.addImage(incremental.image, imgPath)
-        archive.addMetadata(incremental.imgMetadata, metadataPath)
-
-    else:
-        raise ValidationError("No valid archive path for image and no override \
-                               specified, can't append")
-
-
-def bidsToBidsinc(archive: BidsArchive,
-                  subjectID: str,
-                  sessionID: str,
-                  dataType: str,
-                  imageIndex: int = 0,
-                  taskName: str = "",
-                  contrastLabel: str = "",
-                  otherLabels: dict = {}):
-    """
-    Creates a BIDS-Incremental file from the specified part of the BIDS Archive.
-
-    Args:
-        archive: The archive to pull data from
-        subjectID: Subject ID to pull data for (for "sub-control01", ID is
-            "control01")
-        sessionID: Session ID to pull data for (for "ses-2020", ID is "2020")
-        taskName: Task to pull data for (for "task-nback", name is "nback")
-        contrastLabel: Image contrast to pull images for (bold, cbv, or phase)
-        dataType: Type of data to pull (common types: anat, func, dwi, fmap).
-            This string must be the same as the name of the directory containing
-            the image data.
-        otherLabels: Other labels specifying appropriate file to pull data for
-            (for "sub-control01_task-nback_bold", label is "task-nback_bold").
-        imageIndex: Index of image to pull if specifying a 4-D sequence.
-
-    Returns:
-        BIDS-Incremental file with the specified image of the archive and its
-            associated metadata
-
-    Examples:
-        bidsToBidsInc(archive, "01", "2020", "func", "task-nback_bold", 0) will
-        extract the first image of the volume at:
-        "sub-01/ses-2020/func/sub-01_task-nback_bold.nii"
-
-    """
-    archivePath = "sub-{}".format(subjectID)
-    if sessionID is not None:
-        archivePath = os.path.join(archivePath, "ses-{}".format(sessionID))
-    archivePath = os.path.join(archivePath, dataType)
-
-    matchingFilePaths = archive.getFilesForPath(archivePath)
-    niftiPaths = [path for path in matchingFilePaths
-                  if bidsUtils.isNiftiPath(path)]
-    metaPaths = [path for path in matchingFilePaths
-                 if bidsUtils.isJsonPath(path)]
-
-    # Fail if no images
-    if not niftiPaths:
-        logger.debug("Failed to find any matching images in the archive to \
-                      make a BIDS-I from!")
-        return None
-
-    # Warn if no metadata
-    if not metaPaths:
-        logger.debug("Failed to find any matching metadata in the archive to \
-                      include in a BIDS-I!")
-
-    metadata = {'sub': subjectID, 'ses': sessionID, 'task': taskName,
-                'contrast_label': contrastLabel}
-    image = None
-
-    def pathEntitiesMatch(path) -> bool:
-        """
-        Return true if the BIDS entities contained in the file at the given path
-        match the entities provided to the BIDS -> BIDS-I conversion method.
-        """
-        entities = parse_file_entities(path)
-
-        if otherLabels:
-            for label, value in otherLabels.items():
-                if entities.get(label) != value:
-                    return False
-
-        return (entities.get("task") == taskName and
-                entities.get("suffix") == contrastLabel)
-
-    for path in niftiPaths:
-        if pathEntitiesMatch(path):
-            image = readNifti(path)
-            break
-
-    for path in metaPaths:
-        if pathEntitiesMatch(path):
-            with open(path, 'r', encoding='utf-8') as metadataFile:
-                metadata.update(json.load(metadataFile))
-            break
-
-    if image is not None:
-        # https://nipy.org/nibabel/images_and_memory.html
-        targetSlice = image.dataobj[..., imageIndex]
-        newImage = nib.Nifti1Image(targetSlice, image.affine, image.header)
-        newImage.update_header()  # reset header to reflect new dimensions
-        return BidsIncremental(newImage, metadata)
-    else:
-        logger.debug("Failed to find image for BIDS-I in the BIDS archive!")
-        return None
+        # Write out other required files (eg README, dataset description)
